好友列表分页

+++ src/main/java/cn/mifan123/refill/controller/FriendshipController.java
@@ -7,15 +7,15 @@ import cn.mifan123.refill.common.constant.Constants;
 import cn.mifan123.refill.common.exception.BusinessException;
 import cn.mifan123.refill.common.vo.FriendId;
 import cn.mifan123.refill.common.vo.Friendship;
+import cn.mifan123.refill.common.vo.FriendshipIntimacy;
 import cn.mifan123.refill.service.FriendshipService;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.*;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.annotations.ApiIgnore;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletResponse;
 import java.util.Date;
 import java.util.List;
 import java.util.Objects;
@@ -28,12 +28,21 @@ public class FriendshipController {
     @Resource
     private FriendshipService friendshipService;
 
+//    @ApiResponses({
+//            @ApiResponse(code= HttpServletResponse.SC_OK, message = "成功", responseHeaders = {
+//                    @ResponseHeader(name="Total",description="总数据条数")
+//            })
+//    })
     @Auth
     @ApiImplicitParam(value="令牌", paramType = "header", required = true, name = Constants.TOKEN_HEADER_NAME, dataType = "String")
     @ApiOperation("获取用户好友列表")
     @GetMapping("")
-    public List<Friendship> getFriendships(@ApiIgnore @CurrentUser Integer userId) {
-        return friendshipService.findAllByUserId(userId, null, null);
+    public List<Friendship> getFriendships(@ApiIgnore @CurrentUser Integer userId,
+                                           @ApiParam("页码，从0开始")@RequestParam(required = false) Integer page,
+                                           @ApiParam("页大小")@RequestParam(required = false) Integer size,
+                                           HttpServletResponse response) {
+        response.setHeader("Total", String.valueOf(friendshipService.count()));
+        return friendshipService.findAllByUserId(userId, page, size);
     }
 
     @Auth
@@ -58,10 +67,31 @@ public class FriendshipController {
     @DeleteMapping("/{id}")
     public void deleteFriendships(@ApiIgnore @CurrentUser Integer userId, @PathVariable("id") Integer id) {
         Friendship friendship = friendshipService.findOne(id);
+        if(friendship == null) {
+            throw new BusinessException("不存在的好友关系");
+        }
         if(!Objects.equals(friendship.getUserId(), userId)) {
             throw new BusinessException("不能删除不属于自己的好友关系");
         }
         friendshipService.delete(id);
     }
+
+
+    @Auth
+    @ApiImplicitParam(value="令牌", paramType = "header", required = true, name = Constants.TOKEN_HEADER_NAME, dataType = "String")
+    @ApiOperation("好友互动")
+    @PutMapping("/{id}/")
+    public void friendshipsInteractive(@ApiIgnore @CurrentUser Integer userId, @PathVariable("id") Integer id,
+                                       @RequestBody FriendshipIntimacy friendshipIntimacy) {
+        Friendship friendship = friendshipService.findOne(id);
+        if(friendship == null) {
+            throw new BusinessException("不存在的好友关系");
+        }
+        if(!Objects.equals(friendship.getUserId(), userId)) {
+            throw new BusinessException("不属于自己的好友关系");
+        }
+        friendshipService.save(friendship);
+    }
+
 
 }
